+++ ton_client/src/net/types.rs
@@ -38,7 +38,7 @@ pub fn default_out_of_sync_threshold() -> u32 {
     15000
 }
 
-pub fn default_max_reconnect_timeout() -> u32 {
+pub fn default_reconnect_timeout() -> u32 {
     120000
 }
 
@@ -72,10 +72,10 @@ fn deserialize_out_of_sync_threshold<'de, D: Deserializer<'de>>(
     Ok(Option::deserialize(deserializer)?.unwrap_or(default_out_of_sync_threshold()))
 }
 
-fn deserialize_max_reconnect_timeout<'de, D: Deserializer<'de>>(
+fn deserialize_reconnect_timeout<'de, D: Deserializer<'de>>(
     deserializer: D,
 ) -> Result<u32, D::Error> {
-    Ok(Option::deserialize(deserializer)?.unwrap_or(default_max_reconnect_timeout()))
+    Ok(Option::deserialize(deserializer)?.unwrap_or(default_reconnect_timeout()))
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, ApiType)]
@@ -88,15 +88,10 @@ pub struct NetworkConfig {
     /// This parameter is prevailing over `server_address`.
     pub endpoints: Option<Vec<String>>,
 
-    /// Deprecated. You must use `network.max_reconnect_timeout` that allows to specify maximum network resolving timeout.
+    /// Deprecated. You must use `network.reconnect_timeout` that allows to specify maximum network resolving timeout.
     #[serde(default = "default_network_retries_count",
     deserialize_with = "deserialize_network_retries_count")]
     pub network_retries_count: i8,
-
-    /// Maximum time for sequential reconnections in ms. Default value is 120000 (2 min)
-    #[serde(default = "default_max_reconnect_timeout",
-    deserialize_with = "deserialize_max_reconnect_timeout")]
-    pub max_reconnect_timeout: u32,
 
     /// The number of automatic message processing retries that SDK performs
     /// in case of `Message Expired (507)` error - but only for those messages which 
@@ -128,6 +123,11 @@ pub struct NetworkConfig {
     deserialize_with = "deserialize_out_of_sync_threshold")]
     pub out_of_sync_threshold: u32,
 
+    /// Maximum time for sequential reconnections in ms. Default value is 120000 (2 min)
+    #[serde(default = "default_reconnect_timeout",
+    deserialize_with = "deserialize_reconnect_timeout")]
+    pub reconnect_timeout: u32,
+
     /// Access key to GraphQL API. At the moment is not used in production
     pub access_key: Option<String>,
 }
@@ -138,7 +138,7 @@ impl Default for NetworkConfig {
             server_address: None,
             endpoints: None,
             network_retries_count: default_network_retries_count(),
-            max_reconnect_timeout: default_max_reconnect_timeout(),
+            reconnect_timeout: default_reconnect_timeout(),
             message_retries_count: default_message_retries_count(),
             message_processing_timeout: default_message_processing_timeout(),
             wait_for_timeout: default_wait_for_timeout(),